+++ src/cloudai/workloads/nccl_test/nccl_comparison_report.py
@@ -21,16 +21,27 @@ from pathlib import Path
 import cloudai.metrics
 from cloudai.core import System, TestScenario
 from cloudai.report_generator.comparison_report import (
-    ComparisonReport,
     ComparisonReportConfig,
-    ComparisonSection,
 )
-from cloudai.report_generator.groups import GroupedTestRuns
+from cloudai.report_generator.sol_comparison_report import SOLComparisonReport
 from cloudai.workloads.nccl_test.nccl import NCCLTestDefinition
 
 
-class NcclComparisonReport(ComparisonReport):
+class NcclComparisonReport(SOLComparisonReport):
     """Comparison report for NCCL Test."""
+
+    metric_views = (
+        cloudai.metrics.MetricView(
+            cloudai.metrics.LATENCY,
+            x_dimension=cloudai.metrics.SIZE_BYTES.key,
+            series_dimensions=(cloudai.metrics.PLACEMENT.key,),
+        ),
+        cloudai.metrics.MetricView(
+            cloudai.metrics.BANDWIDTH,
+            x_dimension=cloudai.metrics.SIZE_BYTES.key,
+            series_dimensions=(cloudai.metrics.PLACEMENT.key,),
+        ),
+    )
 
     def __init__(
         self, system: System, test_scenario: TestScenario, results_root: Path, config: ComparisonReportConfig
@@ -41,6 +52,3 @@ class NcclComparisonReport(ComparisonReport):
     def load_test_runs(self):
         super().load_test_runs()
         self.trs = [tr for tr in self.trs if isinstance(tr.test, NCCLTestDefinition)]
-
-    def build_sections(self, cmp_groups: list[GroupedTestRuns]) -> list[ComparisonSection]:
-        return self.build_metric_sections(cmp_groups, (cloudai.metrics.LATENCY, cloudai.metrics.BANDWIDTH))
