+++ src/cloudai/workloads/nixl_bench/nixl_summary_report.py
@@ -21,17 +21,28 @@ from pathlib import Path
 import cloudai.metrics
 from cloudai.core import System, TestScenario
 from cloudai.report_generator.comparison_report import (
-    ComparisonReport,
     ComparisonReportConfig,
-    ComparisonSection,
 )
-from cloudai.report_generator.groups import GroupedTestRuns
+from cloudai.report_generator.sol_comparison_report import SOLComparisonReport
 
 from .nixl_bench import NIXLBenchTestDefinition
 
 
-class NIXLBenchComparisonReport(ComparisonReport):
+class NIXLBenchComparisonReport(SOLComparisonReport):
     """Comparison report for NIXL Bench."""
+
+    metric_views = (
+        cloudai.metrics.MetricView(
+            cloudai.metrics.LATENCY,
+            x_dimension=cloudai.metrics.SIZE_BYTES.key,
+            series_dimensions=(),
+        ),
+        cloudai.metrics.MetricView(
+            cloudai.metrics.BANDWIDTH,
+            x_dimension=cloudai.metrics.SIZE_BYTES.key,
+            series_dimensions=(),
+        ),
+    )
 
     def __init__(
         self, system: System, test_scenario: TestScenario, results_root: Path, config: ComparisonReportConfig
@@ -42,6 +53,3 @@ class NIXLBenchComparisonReport(ComparisonReport):
     def load_test_runs(self):
         super().load_test_runs()
         self.trs = [tr for tr in self.trs if isinstance(tr.test, NIXLBenchTestDefinition)]
-
-    def build_sections(self, cmp_groups: list[GroupedTestRuns]) -> list[ComparisonSection]:
-        return self.build_metric_sections(cmp_groups, (cloudai.metrics.LATENCY, cloudai.metrics.BANDWIDTH))
